release 1.4

+++ website/cart/templatetags/cart_tag.py
@@ -1,8 +1,11 @@
 from django import template
 
+
 register = template.Library()
 
 
 @register.filter()
 def multiply(value, arg):
-    return float(value) * arg
+    return int(value) * arg
+
+
